Moved Router from index to App, refactoring
