+++ src/Components/App/App.js
@@ -1,11 +1,14 @@
-import { Switch, Route, Redirect } from "react-router-dom";
+import { Switch, Router, Route, Redirect } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { createBrowserHistory } from "history";
 
 import "./App.scss";
 
 import Header from "../../Shared/Header/Header";
 import PersonTable from "../PersonTable/PersonTable";
 import PersonDetails from "../PersonDetails/PersonDetails";
+
+export const history = createBrowserHistory();
 
 function App() {
   const person = useSelector(state => state.person);
@@ -15,14 +18,18 @@ function App() {
       <Header headerText={"Address Book App"} />
 
       <main className="main">
-        <Switch>
-          <Route exact path="/" component={PersonTable} />
-          <Route
-            path="/details"
-            render={() =>
-              person.firstName !== "" ? <PersonDetails /> : <Redirect to="/" />}
-          />
-        </Switch>
+        <Router history={history}>
+          <Switch>
+            <Route exact path="/" component={PersonTable} />
+            <Route
+              path="/details"
+              render={() =>
+                person.firstName !== ""
+                  ? <PersonDetails />
+                  : <Redirect to="/" />}
+            />
+          </Switch>
+        </Router>
       </main>
     </div>
   );
