+++ src/index.js
@@ -3,8 +3,6 @@ import ReactDOM from "react-dom";
 import { createStore, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
 import { Provider } from "react-redux";
-import { Router } from "react-router-dom";
-import { createBrowserHistory } from "history";
 
 import "./index.scss";
 import App from "./Components/App/App";
@@ -19,14 +17,10 @@ const store = createStore(
   applyMiddleware(thunk)
 );
 
-export const history = createBrowserHistory();
-
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <Router history={history}>
-        <App />
-      </Router>
+      <App />
     </Provider>
   </React.StrictMode>,
   document.getElementById("root")
